+++ include/ionize.h
@@ -24,25 +24,25 @@
 struct ionize { 
     int nR_Dens;
     int nZ_Dens;
-    double* DensGridr;
-    double* DensGridz;
-    double* ne;
+    float* DensGridr;
+    float* DensGridz;
+    float* ne;
     int nR_Temp;
     int nZ_Temp;
-    double* TempGridr;
-    double* TempGridz;
-    double* te;
+    float* TempGridr;
+    float* TempGridz;
+    float* te;
     int nTemperaturesIonize;
     int nDensitiesIonize;
-    double* gridDensity_Ionization;
-    double* gridTemperature_Ionization;
-    double* rateCoeff_Ionization;
-    const double dt;
-        ionize(double _dt,int _nR_Dens,int _nZ_Dens,double* _DensGridr,
-    double* _DensGridz,double* _ne,int _nR_Temp, int _nZ_Temp,
-    double* _TempGridr, double* _TempGridz,double* _te,int _nTemperaturesIonize,
-    int _nDensitiesIonize,double* _gridTemperature_Ionization,double* _gridDensity_Ionization,
-    double* _rateCoeff_Ionization
+    float* gridDensity_Ionization;
+    float* gridTemperature_Ionization;
+    float* rateCoeff_Ionization;
+    const float dt;
+        ionize(float _dt,int _nR_Dens,int _nZ_Dens,float* _DensGridr,
+    float* _DensGridz,float* _ne,int _nR_Temp, int _nZ_Temp,
+    float* _TempGridr, float* _TempGridz,float* _te,int _nTemperaturesIonize,
+    int _nDensitiesIonize,float* _gridTemperature_Ionization,float* _gridDensity_Ionization,
+    float* _rateCoeff_Ionization
               ) : dt(_dt), nR_Dens(_nR_Dens), nZ_Dens(_nZ_Dens), DensGridr(_DensGridr), DensGridz(_DensGridz),ne(_ne),
         nR_Temp(_nR_Temp), nZ_Temp(_nZ_Temp), TempGridr(_TempGridr), TempGridz(_TempGridz), te(_te),
    nTemperaturesIonize(_nTemperaturesIonize), nDensitiesIonize(_nDensitiesIonize), gridTemperature_Ionization(_gridTemperature_Ionization), gridDensity_Ionization(_gridDensity_Ionization), rateCoeff_Ionization(_rateCoeff_Ionization) {}
@@ -50,14 +50,16 @@ struct ionize {
         CUDA_CALLABLE_MEMBER_DEVICE 
                 void operator()(Particle &p) const { 
 	if(p.hitWall == 0.0){        
-	double tion = interpRateCoeff2d ( p.charge, p.x, p.y, p.z,nR_Temp,nZ_Temp, TempGridr,TempGridz,te,DensGridr,DensGridz, ne,nTemperaturesIonize,nDensitiesIonize,gridTemperature_Ionization,gridDensity_Ionization,rateCoeff_Ionization );	
-    double P1 = 1-exp(-dt/tion);
+        //std::cout << "interpolating rate coeff at "<< p.x << " " << p.y << " " << p.z << std::endl;
+        float tion = interpRateCoeff2d ( p.charge, p.x, p.y, p.z,nR_Temp,nZ_Temp, TempGridr,TempGridz,te,DensGridr,DensGridz, ne,nTemperaturesIonize,nDensitiesIonize,gridTemperature_Ionization,gridDensity_Ionization,rateCoeff_Ionization );	
+    float P1 = 1-exp(-dt/tion);
 	
+        //std::cout << "calculating r1 " << std::endl;
 	#ifdef __CUDACC__
-	double r1 = curand_uniform(&p.streams[0]);
+	float r1 = curand_uniform(&p.streams[0]);
 	#else
-	std::uniform_real_distribution<double> dist(0.0, 1.0);
-	double r1=dist(p.streams[0]);
+	std::uniform_real_distribution<float> dist(0.0, 1.0);
+	float r1=dist(p.streams[0]);
 	#endif
 
 	if(r1 <= P1)
